refactor(react-scheduling): split AppointmentFinder.utils along its seams
`AppointmentFinder.utils.ts` had grown to 1,269 lines that twelve files import
from, which made every component depend on everything the hub touched: a card
that only formats a time pulled in `$find` paging and the scheduling-parameter
readers along with it.

Split it into six modules, layered so nothing cycles. `roles` is the root and
depends on nothing; `times` and `schedules` sit above it; `find` and `assemble`
above those. `params` is kept apart because it is the only part reaching for
`getScheduleParameters` and `durationToMinutes`, so the other five need nothing
outside what is already released.

No behavior change: the tests are the same tests, moved to sit beside the module
they cover.
