+++ packages/react-scheduling/src/index.ts
@@ -11,7 +11,12 @@ export * from './AppointmentFinder/AppointmentCriteriaForm';
 export * from './AppointmentFinder/AppointmentCustomTimeCard';
 export * from './AppointmentFinder/AppointmentDayTimes';
 export * from './AppointmentFinder/AppointmentFinder';
-export * from './AppointmentFinder/AppointmentFinder.utils';
+export * from './AppointmentFinder/AppointmentFinder.assemble';
+export * from './AppointmentFinder/AppointmentFinder.find';
+export * from './AppointmentFinder/AppointmentFinder.params';
+export * from './AppointmentFinder/AppointmentFinder.roles';
+export * from './AppointmentFinder/AppointmentFinder.schedules';
+export * from './AppointmentFinder/AppointmentFinder.times';
 export * from './AppointmentFinder/AppointmentLocationSelect';
 export * from './AppointmentFinder/AppointmentPatientSelect';
 export * from './AppointmentFinder/AppointmentPickList';
